Analytics event interruption
