--- src/lib/analytics/analytics-event.enum.ts
@@ -1,16 +1,0 @@
-export enum AnalyticsEventCategory {
-  General = 'General',
-  ButtonPress = 'ButtonPress',
-  FormChange = 'FormChange',
-  FormSubmit = 'FormSubmit',
-  FormSubmitSuccess = 'FormSubmitSuccess',
-  FormSubmitFail = 'FormSubmitFail',
-  PageOpened = 'PageOpened'
-}
-
-export enum AnalyticsEventEnum {
-  AnalyticsEnabled = 'AnalyticsEnabled',
-  AnalyticsDisabled = 'AnalyticsDisabled',
-  LanguageChanged = 'LanguageChanged',
-  FiatCurrencyChanged = 'FiatCurrencyChanged'
-}
